requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
cfgv==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.4.0
Django==4.2.24
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
identify==2.6.14
kombu==5.5.4
Markdown==3.9
nodeenv==1.9.1
packaging==25.0
platformdirs==4.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
